renamed tabs to match other component structure, and added first section
