+++ components/_tab-tile/tabs-tile-with-icons.hbs
@@ -1,0 +1,52 @@
+<div class="msds-tabs">
+    <div class="msds-tabs__container">
+        <div class="msds-tabs__spacer"></div>
+
+        <div class="msds-tabs__tabs-container container">
+
+            <div class="msds-tabs__tab msds-tabs__tab--active" data-index="0" tabindex="1">
+                <div class="msds-tabs__tab-svg msds-download-illustration" title="Download Illustration"></div>
+                <div class="msds-tabs__tab-text">
+                    placeholderHeading
+                </div>
+            </div>
+
+            <div class="msds-tabs__tab" data-index="1" tabindex="1">
+                <div class="msds-tabs__tab-svg msds-reliable" title="Reliable"></div>
+                <div class="msds-tabs__tab-text">
+                    placeholderHeading
+                </div>
+            </div>
+
+            <div class="msds-tabs__tab" data-index="2" tabindex="1">
+                <div class="msds-tabs__tab-svg msds-scalable" title="Scalable"></div>
+                <div class="msds-tabs__tab-text">
+                    placeholderHeading
+                </div>
+            </div>
+
+            <div class="msds-tabs__tab" data-index="3" tabindex="1">
+                <div class="msds-tabs__tab-svg msds-end-user" title="End User"></div>
+                <div class="msds-tabs__tab-text">
+                    placeholderHeading
+                </div>
+            </div>
+
+            <div class="msds-tabs__tab" data-index="4" tabindex="1">
+                <div class="msds-tabs__tab-svg msds-interconnect" title="Interconnect"></div>
+                <div class="msds-tabs__tab-text">
+                    placeholderHeading
+                </div>
+            </div>
+
+            <div class="msds-tabs__tab" data-index="5" tabindex="1">
+                <div class="msds-tabs__tab-svg msds-xprotect" title="XProtect"></div>
+                <div class="msds-tabs__tab-text">
+                    placeholderHeading
+                </div>
+            </div>
+        </div>
+
+        <div class="msds-tabs__spacer"></div>
+    </div>
+</div>
