+++ src/js/main.js
@@ -1,5 +1,5 @@
 import '../scss/main.scss'
-import 'components/tab-tile/tab-tile.js'
+import 'components/_tab-tile/tab-tile.js'
 import 'components/_alert/alert.js'
 import 'components/forms/_text-areas/text-area.js'
 import 'components/button-cards/card.js'
