+++ src/scss/main.scss
@@ -15,7 +15,7 @@
 @import 'components/button-cards/_contact-card/contact-card.scss';
 @import 'components/button-cards/_solution-card/solution-card.scss';
 @import 'components/button-cards/_basic-card/basic-card.scss';
-@import 'components/tab-tile/tab-tile.scss';
+@import 'components/_tab-tile/tab-tile.scss';
 @import 'components/forms/_input-fields/input-field.scss';
 @import 'components/forms/_text-areas/text-area.scss';
 @import 'components/forms/_checkboxes/checkbox.scss';
